restructure to different service for restocking
